refactor: extract function in its own file to split logic in different files and avoid massive files
