+++ src/libs/algolia/fetchAlgolia/fetchVenuesModules.ts
@@ -1,12 +1,8 @@
 import { Venue, VenuesModuleParameters } from 'features/home/types'
-import { LocationType } from 'features/search/enums'
-import { AlgoliaVenue, FiltersArray } from 'libs/algolia'
-import { VenuesFacets } from 'libs/algolia/enums'
+import { AlgoliaVenue } from 'libs/algolia'
 import { captureAlgoliaError } from 'libs/algolia/fetchAlgolia/AlgoliaError'
-import { buildGeolocationParameter } from 'libs/algolia/fetchAlgolia/buildAlgoliaParameters/buildGeolocationParameter'
-import { getVenueTypeFacetFilters } from 'libs/algolia/fetchAlgolia/buildAlgoliaParameters/getVenueTypeFacetFilters'
+import { buildVenuesQueryOptions } from 'libs/algolia/fetchAlgolia/buildVenuesQueryOptions'
 import { client } from 'libs/algolia/fetchAlgolia/clients'
-import { adaptGeolocationParameters } from 'libs/algolia/fetchAlgolia/helpers/adaptGeolocationParameters'
 import { buildHitsPerPage } from 'libs/algolia/fetchAlgolia/utils'
 import { env } from 'libs/environment'
 import { Position } from 'libs/geolocation'
@@ -37,43 +33,6 @@ export const fetchVenuesModules = async (
     return [] as Venue[][]
   }
 }
-
-export const buildVenuesQueryOptions = (params: VenuesModuleParameters, userLocation: Position) => {
-  const { aroundRadius, isGeolocated, tags = [], venueTypes = [] } = params
-
-  const locationFilter = adaptGeolocationParameters(userLocation, isGeolocated, aroundRadius) ?? {
-    locationType: LocationType.EVERYWHERE,
-  }
-
-  const facetFilters: FiltersArray = []
-
-  if (tags.length) {
-    const tagsPredicate = buildTagsPredicate(tags)
-    facetFilters.push(tagsPredicate)
-  }
-
-  if (venueTypes.length) {
-    const venueTypesPredicate = buildVenueTypesPredicate(venueTypes.map(getVenueTypeFacetFilters))
-    facetFilters.push(venueTypesPredicate)
-  }
-
-  // We want to show on home page only venues that have at least one offer that is searchable in algolia
-  const hasAtLeastOneBookableOfferPredicate = [
-    `${VenuesFacets.has_at_least_one_bookable_offer}:true`,
-  ]
-  facetFilters.push(hasAtLeastOneBookableOfferPredicate)
-
-  return {
-    ...buildGeolocationParameter(locationFilter, userLocation),
-    ...(facetFilters.length > 0 ? { facetFilters } : {}),
-  }
-}
-
-const buildVenueTypesPredicate = (venueTypes: string[]): string[] =>
-  venueTypes.map((venueType) => `${VenuesFacets.venue_type}:${venueType}`)
-
-const buildTagsPredicate = (tags: string[]): string[] =>
-  tags.map((tag: string) => `${VenuesFacets.tags}:${tag}`)
 
 const buildVenue = (venue: AlgoliaVenue): Venue => {
   const socialMedias: Record<string, string> = {}
